- Using `bulma-accordion` package for the institutions panel "navigation". - Replaced the `@fortawesome/fontawesome` with `font-awesome`. - Renamed `resources/assets/js/institutions-panel.vue` to `resources/assets/js/institutions-panel.vue`. - Renamed `resources/assets/js/institutions-pane-institution.vue` to `resources/assets/js/institutions-panel-institution.vue`. - Renamed `resources/assets/js/institutions-pane-institution-account.vue` to `resources/assets/js/institutions-panel-institution-account.vue`.
