--- resources/assets/js/components/institutions-pane-institution-account.vue
@@ -1,54 +1,0 @@
-<template>
-    <li v-bind:id="'account-'+this.accountId">
-        <a class="tooltip is-tooltip-right is-tooltip-multiline" v-bind:data-tooltip="accountTypeTooltipList">
-            <span
-                v-text="this.accountName"
-                v-bind:class="{'badge is-badge-small is-badge-info is-badge-outlined': isAccountTotalVisable}"
-                v-bind:data-badge="'$'+parseFloat(accountTotal).toFixed(2)"
-            ></span>
-        </a>
-    </li>
-</template>
-
-<script>
-    import {Accounts} from "../accounts";
-
-    export default {
-        name: "institutions-pane-institution-account",
-        props: ['id', 'name', 'total'],
-        computed: {
-            accountId: function(){
-                return this.id;
-            },
-            accountName: function(){
-                return this.name;
-            },
-            accountTotal: function(){
-                return this.total;
-            },
-            isAccountTotalVisable: function(){
-                return !isNaN(this.accountTotal);
-            },
-            accountTypeTooltipList: function(){
-                let accountTypes = new Accounts().getAccountTypes(this.accountId);
-                let tooltipList = "";
-                accountTypes.forEach(function(accountType){
-                    tooltipList += "- "+accountType.name+" ("+accountType.last_digits+")\n"
-                });
-                return tooltipList;
-            },
-        },
-        methods: {
-
-        }
-    }
-</script>
-
-<style scoped>
-    li{
-        padding-right: 65px;
-    }
-    .badge::after{
-        margin: 6px 10px 0;
-    }
-</style>
